close #96
Change COUCHDB_DATABASES so that password is hidden in DEBUG

Changes the COUCHDB_DATBASES tuple to be more like the django 1.3 dict that contains the normal database connection. What this allows is the password to be hidden in debug fields as well as being more inline with the way django does setup for its databases. This is only a user interface change as the change just puts it all back together when its needed.

Initial patch from @mlakewood, improved by @georgemarshall